+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ComponentDict.global_observers` for observing changes in components globally.
+
 ### Changed
 - `ComponentDict` internals have changed.  No longer uses attrs and has a less complex constructor.
   Should still be able to unpickle plain instances pickled in v1.